requirements-parquet.txt
pyarrow
pandas
Pillow